fix: run service installer off the UI thread
